requirements.txt
flask==3.1.2
gunicorn==23.0.0
alembic==1.16.5
sqlalchemy==2.0.43
flask-sqlalchemy==3.1.1
flask-login==0.6.3
authlib==1.6.4
python-dotenv==1.1.1
PyMySQL==1.1.2